requirements.txt
Django==2.0.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
